Output SATS in integer format.

+++ src/ch/luethi/rw5tocsv/Vrec.java
@@ -56,7 +56,7 @@ public class Vrec {
 
     // these value are additions fot the text output
     protected int numberOfMeasurements;
-    protected float sats, satsMin, satsMax;
+    protected int sats, satsMin, satsMax;
     protected String date;
     protected String time;
     protected State state;
